newlog: restructuring files in the package
This commit restructures the `newlog` package to improve organization
and maintainability. We split the `newlogd.go` file into multiple files,
each handling a specific aspect of the logging functionality:
- retrieving kernel messages
- retrieving syslog messages
- retrieving memlog messages
- moving and compressing files
- checking quotas
- saving panic files

The tests were split accordingly to match the new structure. This
also provides better visibility over the test coverage.
