+++ pkg/newlog/cmd/newlogd_test.go
@@ -17,25 +17,6 @@ func init() {
 	logger, log = agentlog.Init(agentName)
 }
 
-func TestGzipParsing(t *testing.T) {
-	t.Parallel()
-	g := gomega.NewWithT(t)
-
-	// Test the gzip parsing function
-	keepSentDir = "../testdata/keepSentQueue"
-	oldestLogEntry, err := getOldestLog()
-
-	g.Expect(err).To(gomega.BeNil())
-	g.Expect(oldestLogEntry).NotTo(gomega.BeNil())
-	t.Logf("latestLogEntry: %v\n", oldestLogEntry)
-
-	g.Expect(oldestLogEntry.Content).To(gomega.Equal("memlogd started"))
-
-	logmetrics.OldestSavedDeviceLog = time.Unix(
-		oldestLogEntry.Timestamp.Seconds, int64(oldestLogEntry.Timestamp.Nanos))
-	t.Log("OldestSavedDeviceLog: ", logmetrics.OldestSavedDeviceLog)
-}
-
 func TestGetTimestampFromGzipName(t *testing.T) {
 	t.Parallel()
 	g := gomega.NewWithT(t)
@@ -58,119 +39,5 @@ func TestGetTimestampFromGzipName(t *testing.T) {
 		timestamp, err := types.GetTimestampFromGzipName(file.Name())
 		g.Expect(err).To(gomega.BeNil())
 		g.Expect(timestamp).To(gomega.Equal(comparisonMap[file.Name()]))
-	}
-}
-
-func TestFindMovePrevLogFiles(t *testing.T) {
-	t.Parallel()
-	g := gomega.NewWithT(t)
-
-	collectDir = "../testdata/collect"
-
-	disableLogsForApp := true
-	appD := appDomain{
-		disableLogs: disableLogsForApp,
-	}
-	domainUUID.Store("8ce1cc69-e1bb-4fe3-9613-e3eb1c5f5c4d", appD)
-
-	movefileChan := make(chan fileChanInfo, 5)
-	findMovePrevLogFiles(movefileChan)
-
-	g.Eventually(movefileChan).Should(gomega.HaveLen(3))
-
-	var file1, file2, file3 fileChanInfo
-	select {
-	case file1 = <-movefileChan:
-	case <-time.After(time.Second):
-		t.Fatal("Expected fileChanInfo not received")
-	}
-
-	select {
-	case file2 = <-movefileChan:
-	case <-time.After(time.Second):
-		t.Fatal("Expected fileChanInfo not received")
-	}
-
-	select {
-	case file3 = <-movefileChan:
-	case <-time.After(time.Second):
-		t.Fatal("Expected fileChanInfo not received")
-	}
-
-	files := []fileChanInfo{file1, file2, file3}
-	expectedFiles := map[string]bool{
-		"../testdata/collect/dev.log.keep.397634723":                                     true,
-		"../testdata/collect/dev.log.upload.815590695":                                   false,
-		"../testdata/collect/app.8ce1cc69-e1bb-4fe3-9613-e3eb1c5f5c4d.log.1731935033496": disableLogsForApp,
-	}
-
-	for _, file := range files {
-		expectedNotUpload, exists := expectedFiles[file.tmpfile]
-		g.Expect(exists).To(gomega.BeTrue())
-		g.Expect(file.notUpload).To(gomega.Equal(expectedNotUpload))
-
-		getFileInfo(file)
-	}
-}
-
-func TestGetFileInfo(t *testing.T) {
-	t.Parallel()
-	g := gomega.NewWithT(t)
-
-	tests := []struct {
-		name          string
-		fileChanInfo  fileChanInfo
-		expectedDir   string
-		expectedAppID string
-	}{
-		{
-			name: "Device log file for upload",
-			fileChanInfo: fileChanInfo{
-				tmpfile:   "../testdata/collect/dev.log.upload.123456",
-				isApp:     false,
-				notUpload: false,
-			},
-			expectedDir:   uploadDevDir,
-			expectedAppID: "",
-		},
-		{
-			name: "Device log file to keep",
-			fileChanInfo: fileChanInfo{
-				tmpfile:   "../testdata/collect/dev.log.keep.123456",
-				isApp:     false,
-				notUpload: true,
-			},
-			expectedDir:   keepSentDir,
-			expectedAppID: "",
-		},
-		{
-			name: "App log file for upload",
-			fileChanInfo: fileChanInfo{
-				tmpfile:   "../testdata/collect/app.8ce1cc69-e1bb-4fe3-9613-e3eb1c5f5c4d.log.123456",
-				isApp:     true,
-				notUpload: false,
-			},
-			expectedDir:   uploadAppDir,
-			expectedAppID: "8ce1cc69-e1bb-4fe3-9613-e3eb1c5f5c4d",
-		},
-		{
-			name: "App log file to keep",
-			fileChanInfo: fileChanInfo{
-				tmpfile:   "../testdata/collect/app.8ce1cc69-e1bb-4fe3-9613-e3eb1c5f5c4d.log.123456",
-				isApp:     true,
-				notUpload: true,
-			},
-			expectedDir:   keepSentDir,
-			expectedAppID: "8ce1cc69-e1bb-4fe3-9613-e3eb1c5f5c4d",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			t.Parallel()
-			dirName, appuuid := getFileInfo(tt.fileChanInfo)
-			g.Expect(dirName).To(gomega.Equal(tt.expectedDir))
-			g.Expect(appuuid).To(gomega.Equal(tt.expectedAppID))
-		})
 	}
 }
